Add Redis engine
Added a Redis engine as a storage mechanism. This does add a dependency on the redis gem which is questionable. I was unable to find a way to conditionally load it if a user wanted that engine. Perhaps the engines that depend on external gems should be external gems themselves.